backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
cryptography==41.0.7
reportlab==4.0.7
openpyxl==3.1.2
python-decouple==3.8